refactor: move shared utilites to one module
